♻️ e2e: 各リンクのナビゲーションテストをリファクタリングする
